Fix loop_mw + newnet

+++ include/private/misc.h
@@ -43,7 +43,7 @@ extern void			*gl_loop_data;
 
 extern t_bunny_threadpool	*gl_bunny_loop_threadpool;
 
-t_bunny_response		network_event(unsigned int		v,
+t_bunny_response		network_event(double			v,
 					      void			*dat);
 
 void				_real_call(const t_bunny_prototype	*function,
